requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=5.4.0
tqdm>=4.62.0

# AWS dependencies
boto3>=1.18.0
awswrangler>=2.11.0
botocore>=1.21.0

# Machine Learning
scikit-learn>=0.24.0
xgboost>=1.4.0
tensorflow>=2.6.0
torch>=1.9.0

# Data processing
pyarrow>=5.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Time series specific (yochlol)
sktime>=0.8.0
tslearn>=0.5.0

# Utilities
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0

# Optional but recommended
plotly>=5.3.0
dash>=2.0.0
streamlit>=0.87.0